Restructure BarChart
