+++ projects/js-packages/charts/src/components/bar-chart/bar-chart.tsx
@@ -22,8 +22,7 @@ import { Legend, useChartLegendItems } from '../legend';
 import { SingleChartContext } from '../single-chart-context';
 import { AccessibleTooltip, useKeyboardNavigation } from '../tooltip/accessible-tooltip';
 import { withResponsive } from '../with-responsive';
-import styles from './bar-chart.module.scss';
-import { useBarChartOptions } from './use-bar-chart-options';
+import { useBarChartOptions, styles } from './private';
 import type { BaseChartProps, DataPointDate, SeriesData, Optional } from '../../types';
 import type { ResponsiveConfig } from '../with-responsive';
 import type { RenderTooltipParams } from '@visx/xychart/lib/components/Tooltip';
