+++ projects/js-packages/charts/src/components/bar-chart/private/bar-chart.module.scss
