+++ projects/js-packages/charts/src/components/bar-chart/private/index.ts
@@ -1,0 +1,2 @@
+export { useBarChartOptions } from './use-bar-chart-options';
+export { default as styles } from './bar-chart.module.scss';
